Public SDK: Build using a special option to normal makefile, refs #5781
Before the Public SDK was built using a separate package.
Now that it is possible to choose between several different kernel
versions, this does not work. Instead a new menuconfig option to the
normal bcmkernel package has been added that makes it build the
Public SDK of the current release.

NOTE: Make sure that the kernel version selected in menuconfig
      matches the saved Public SDK tarball of the current release.

Only the latest release (4.16L.03) has been tested, the older version
builds are killed with an eror message if Public SDK is selected.