bcmkernel-4.16L.04: resurrected Bluetooth.
From the long forgotten realm of 4.16L.03 the rotten corps of a hacky
backport was dragged out in the open with powerful magical spells
constructed by the church of gitology.

The reanimation took place during many hours of hard manual labor where
whole body parts was replaced and rearranged according to ancient
manuscripts written in a long forgotten language.

While this has not been blessed with holy penguin pee by our Lord the
mighty Torvalds the complicator of gnu has compiled it to ELF ingots
that has been forged to a coherent whole by the blacksmith of ld.

what could possible go wrong?