New version string used by inteno for u-boot.
We need to have a version that we can change whenever we want to force and
update of u-boot. The u-boot version still exists.

To change the version and force an update change the PKG_RELEASE variable
in package Makefile.

uboot version is now checked on startup and if the installed one is too old
it will be updated.