Make dev-check check for SSH instead of non-HTTP
In order to determine, whether we are developers or not, check the
git-remotes for SSH-URLs instead of checking for the non-existence of
any HTTP-URLs. This fixes dev-mode not being detected properly when
curious devs add other HTTP-remotes such as upstream OpenWRT.

This also makes the check consistent with the check done in the main
"iop" script in the root.